Add Note Activity
Moved the logic for adding notes to an Add Note activity and updated the launch activity to display a FAB with a plus icon that launches the Add Note screen.
