+++ app/src/main/AndroidManifest.xml
@@ -16,5 +16,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".AddNoteActivity"/>
     </application>
 </manifest>
